* commit '42e3abdfd8ff276f6e2333357b62eed3ca2f5f24':
  update node script
  update to support electron-printer
  update to build script to only download electron-prebuild to version 1.2.5
  rename build files
  update script to accept parameters to send to printer module
  update on fix with the script where it only support version 1.2.5
  change to local test file demo

+++ scripts/production.js
@@ -6,6 +6,9 @@ let path = require('path'),
     config = require("../electron.config.js"),
     shell = require('shelljs');
 
+var npmScripts = typeof (process.argv[2]) === 'undefined';
+
+
 const APPLICATION_SRC = path.join(path.dirname(__dirname), config.source);
 
 let electron_printer = path.join(APPLICATION_SRC, 'node_modules', 'electron-printer');
@@ -14,11 +17,12 @@ let command = "\"./node_modules/.bin/electron-packager\" app/",
 //build the command script based on config files
     _c = [
         command
-        , "--platform=win32"
-        , "--arch=ia32"
+        , "--platform=" + config.platform
+        , "--arch=" + config.arch
         , "--asar"
         , "--out=" + config.distribution
         , "--overwrite"
+        , "--version=\"" + config.electronVersion + "\""
     ];
 
 
@@ -37,25 +41,29 @@ if (fs.existsSync(electron_printer)) {
         "--dist-url=https://atom.io/download/atom-shell"
     ].join(" ");
 
-    shell.rm('-rf', path.join(APPLICATION_SRC, 'config.json'));
+    if (npmScripts)shell.rm('-rf', path.join(APPLICATION_SRC, 'config.json'));
     shell.cd(electron_printer);
 
-    if (shell.exec(printer_bin, {silent: true}).code !== 0) {
+    if (shell.exec(printer_bin).code !== 0) {
         console.log('Error: Failed to build electron-printer');
+    } else {
+        console.log('Build electron-printer');
     }
 
 }
 
+if (npmScripts) {
+    fs.writeFile(path.join(APPLICATION_SRC, 'config.json'), JSON.stringify(config), function (err) {
+        if (err) return console.log(err);
+        //back to root
+        shell.cd(path.join(APPLICATION_SRC, '..'));
+        shell.exec((_c.join(" ")));
 
-fs.writeFile(path.join(APPLICATION_SRC, 'config.json'), JSON.stringify(config), function (err) {
-    if (err) return console.log(err);
-    //back to root
-    shell.cd(path.join(APPLICATION_SRC, '..'));
-    shell.exec((_c.join(" ")));
+        //remove the file that we just created
+        shell.rm('-rf', path.join(APPLICATION_SRC, 'config.json'));
 
-    //remove the file that we just created
-    shell.rm('-rf', path.join(APPLICATION_SRC, 'config.json'));
+    });
+}
 
-});
